global: Add destroy method breaking reference cycles through disposition
The shell was destroying the JS Context using a specific method on exit,
however, since the context is now unreffed as part of the disposition, we can
just force the disposition of ShellGlobal triggering the context finalization.

We need to act like this because at this point the global object will
have two references, one is the original instance we constructed and the
other one is owned by the JS context itself. This leads to a reference
cycle, and so we've to break it in this way.

Define _shell_global_destroy() private function where we force the
disposition the global object and we unref it.