environment: Add connect_once to handle one-shot signals using promises
We have a number of places in code where we need to temporary handle a
signal for just an emission, this can now be managed easily using
promises in a generic way, so adding a method to wait for a signal
emission that can be used using the promises syntax.

Once the signal is emitted the promise is resolved, due to its
limitation the signal arguments are passed using the same gjs functions
assumptions: if a single value is returned it's just a value, otherwise
we pass an array of arguments.
The object is not passed as argument as we assume that it can be easily
referenced using arrow-functions if needed.

The promise we return is a bit special as it contains a cancel() method
that will reject it without having to use a GCancellable for that.

This is done by extending a promise so that the then method will return
a promise of the same kind, however we need to ensure that the
properties on the returned object are preserved in order to make the
cancel and other methods to work.