PromiseUtils: Add CancellablePromise, a Promise with a GCancellable attached
Javascript promises don't allow to be cancelled by default and they'd
just run until an error happens.

However there are cases where we want to gracefully cancel the execution
of a Promises chain, so add a Promise implementation that may contain a
GCancellable that is monitored for cancellations.

So if trying to resolve the promise when the cancellable is already
cancelled or when the cancellation happens we just reject the promise
with a G_IO_CANCELLED error so that it can be filtered in try/catches or
catche() callbacks.

The then implementation of the default Promise needs to be overridden in
order to make possible to reach the root promise object we created and
so to be able to cancel it from a child promise.

Adding to a new file that needs to have less imports possible, so that
we can include it early in our environment.