gdm: Replace ConsecutiveBatch with async functions using CancellablePromise
Timed login logic was using some custom code that was built when async
wasn't yet a thing in JS, however this is not any longer the case and we
can replace the logic using Promise-based async functions.

However, since each of the actions we provide need to be cancelled
easily, we replace the batch handler with a cancellable that indicates
us the state of the timed login, while we pass it to the various
CancellablePromise's so that we stop the current request chain once a
cancellation happens.