gdm/loginDialog: Make possible to pass a GCancellable to verifier begin
In order to be able to cancel multiple requests as in the case of the
User selection promise we need to have a single cancellable to stop all
the ongoing requests.

So make possible to provide a GCancellable to the user verifier so that
it can be used to stop all the promises we've batched.

In future reiterations the GCancellable could be replace the verifier
cancel() method usage.