environment: Make ease() functions to be async
Ease functions can be used nested and many times we need an easing to be
completed in order to do an action, currently we rely on callbacks but
this can be easily converted to support Promises as well.

So make the easing functions to always to be async so that they will
return a promise that is going to be resolved by the callback.

The callback will resolve successfully the promise only if the easing
has completed, as this will allow to stop a promises chain easily, while
in case it was stopped we'll return a Cancelled error so that it can be
easily filtered out in catch handlers.

The callback functions are still available and working, but this will
allow to use easing with the promises idiom or in async functions.