system: Do not use Power Off or Log Out labels when they are unavailable
During lock screen or login dialog we show the system menu contains the
submenu "Power Off / Log Out" label even if none of such actions are
available.

In fact, it's never possible to log out at login screen, while it's
not possible to either logout or shut down the system during the lock
screen.

So, set the session label depending on the available actions.