popupMenu: Add PopupAutoSubMenuMenuItem an hybrid submenu and menu item
When submenu contains only a single menu item, it can be just a waste of
clicks to have a submenu for this purpose.

Add then a PopupSubMenuMenuItem extension that acts exactly like a
normal SubMenuItem when there are multiple items, however when only one
item is visible, it will be just like a normal menu, with the label and
icon (if present) bound to the actual visible sub-item and performing
its action.