system: Use a PopupAutoSubMenuMenuItem for the session sub-menu
The session submenu may have a variable number of menu entries, and
in some scenarios (depending on system capabilities and lockdown
options) or during lockscreen only one could be available.

In such scenario, having to select a "Power Off / Log Out" item when
only a "Suspend" submenu may be visible looks a bit disorienting other
than a waste of clicks.
So, in such case let's avoid to show use the submenu at all, and let's
just present the only available menu item.

This also allows to monitor visibility changes on session menu items,
so that we can update the submenu label if needed (i.e. when locking
down the logout option) or hide it when not needed without further
custom code.