system: Use PopuImageMenuItem's for system icons with hidden images
In case we have only one menu item visible we're going to use its label
as the visible submenu item, but we keep the default icon visible.

This may be misleading in some scenarios, so let's reflect it also for
the icons.

We can easily do it using PopupMiamgeMenuItem's with hidden icons
leaving the parent menu item to inherit them if it's the case.