gdm: Make queueMessageTimeout function async
We can now use a timeout promise to wait for the message reading time
and rely on the user verifier cancellable to stop it.

We need to use a local cancellable variable to ensure that we don't end
up using the wrong one if the message has been frozen for a while.