environment: Add cancellable parameter to ease functions
There are situations in which we want to cancel a set of easing actions
happening to multiple actors, while the transitions can be removed per
actor pretty easily, it's not possible to stop a set of them in a single
shot. By having a cancellable that can be shared by multiple transitions
we can control this better.

Per the fact we're internally using cancellable promises we can easily
support this for the easing functions as well.

Added also a shorter path to immediately stop the transition if trying
to start it with a cancelled cancellable, this leads to throwing a
cancelled error as per the callback we define.