croco/stylesheet: Move app_data to private area and add DestroyNotify support
Stylesheet data used to be a simple pointer modifiable by any
application and never touched by libcroco itself.

However it's useful to make the library control its lifecycle together
with the stylesheet itself when a destroy notify function is defined.

So, move the app_data pointer to the private area so that we prevent it
to be modified from applications and add a GDestroyNotify function as well.

As per this define a getter and setter that allows to manage it safely.