croco/stylesheet: Move ref-count to private implementation
This always meant to be a private bit so there's no point to expose it.

We use a private structure type for this without doing extra allocations
as in other croco structures, since it just unneeded to have an extra
pointer just for this, and we can just use a bigger allocation from the
start that is hidden to the library users.

Also adding such type will allow us to add some more bits we don't
want to be directly modifiable from outside.