oVirt/vmware: Do not use a singleton to initialize credential managers
There's no point to keep the credentials manager around when not in
lockscreen or login screen, so we can avoid use a singleton to create
them, and initialize an instance only when needed.

While it's true that oVirt can be initialized multiple times (during
lock screen), the implementation is small enough that there's no problem
in having couple of instances active when screen is locked.