smartcardManager: Dispose the token when removed
When a token that is actually a GDBus proxy is removed we don't need it
anymore but we're still monitoring its dbus properties.

To avoid that we used to use disconnectAll(), but the token is not
implementing the Signals interface being a native GOjbect, so this won't
do anything.

So, instead of just disconnecting the signal we connect to, we can
just safely dispose the object, to ignore gdbus events.