gdm: Only reset the preempting Service after verification-failed event
When the preempting service is set that will be our foreground service.
If a conversation is stopped we're going to mark the verification as
failed and during the callback we may still be able to check whether if
the failed service was the foreground one or not.

So, in order to ensure that when verification-failed is emitted we
preserve the previous state, just unset the preempting service once
we've notified our clients.