authPrompt: Only clear the entry if the failed service was the querying service
As we don't animate if a non-querying service failed, we can just avoid
clearing the interface in such case.

Given that clearing implies also unsetting the default actor in button
well (so hiding the spinner these days), we can just remove this call
instead of repeat it under the same if.