overviewControls: Handle shutdown gracefully
There are two adaptations needed here:

 * Handle the WorkspaceDisplay being destroyed

This happens as a side effect of being a child actor. Handle this by
removing the reference to it, and handling being unmapped more
gracefully by not failing to hide the workspace display if it was
already destroyed.

 * Handle setting the container to 'null'

This is part of being destroyed, so handle it.