promiseUtils: Add SignalConnectionPromiseFull and object.connect_with_promise
This promise is now the base class for SignalConnectionPromise as it can
allow to add more control to the signal handlers that will now be able
to manually decide whether to reject or resolve a promise at callback
time, and this will permit also to return a value to signal accumulators
or to be able to delay the resolution of a promise, depending on the
signal arguments.

Since we're now supporting more flags on connecting, also allow to
support connect_after.

Add tests