accessDialog: Allow requests from apps in STARTING state
Currently we restrict requests to the focus app. That's a bit too
restrictive, for example it prevents Weather from using the Location
portal, because it waits for its state to be ready before showing
its window.

Address this by loosen the restriction a bit and also allow apps
in STARTING state (that is, with an active startup sequence).