dash: Fixes wrong separator position
The separator position takes an active drag placeholder into account, but not
the placeholders and app icons that are currently animating out. As a result,
we may end up with the wrong position and some app icons appear on the wrong
side of the separator.

The merge request !1804 only partially resolves this issue. This is a better
fix.

Fixes #3966, #3799