gdm: Do not run session start fade out effect when not on a VT-capable seat
On a non-VT-capable seat, logind will automatically activate the login
session on creation. On a gnome-shell Wayland session, this causes
animation effects to be stalled indefinitely, causing hang during login
as the animation completion callback is never invoked to launch the
session.

Fix this by checking for current seat's CanTTY property and do not run
session start effect if the current seat is not VT-capable.