extensionSystem: Support ini metadata
JSON is a natural choice for a format that is shared between
desktop JS (gnome-shell) and a web app (extensions.gnome.org),
but it has a downside:

It is hard to integrate with our translation framework, so
there is currently no way to localize user-visible strings
like name and description.

To address this, support reading metadata from a keyfile instead,
which has a well-established way of providing translations.

https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/2288