keyboard: Eliminate wasted space
Make the keyboard stretch to fill all available space. This makes
it possible for the keyboard height to reach 1/3 the display height
even in narrow aspect ratios (i.e. portrait mode), so remove the
maxHeight limitation to allow it to stretch vertically. To avoid
making the keyboard height too small in wide aspect ratios, make
its size independent of the suggestions bar. This will also allow
for hiding the suggestions bar when empty without disturbing the
keyboard size in an upcoming commit.