dnd: Use ClutterGesture to start DnD interactions
With the last commit we introduced ShellDndStartGesture, so now port the
Draggable to actually use that gesture, and to move over to a ClutterGrab
once the DndStartGesture has recognized.

This is shuffling a lot of things around and removing quite a bit of code.

One thing to note is that to allow users of the DndStart gesture to specify
its relationship with other gestures (eg. to allow it run together with a
LongPressGesture), the DndStartGesture is exposed on the draggable as
the startGesture property.