windowManager: Show feedback when latest workspace is reached
Right now when we are at the first/last workspace and we try
to switch via keyboard to the left/right, nothing happens and
only the workspace switcher popup is shown. In order to provide
more feedback that there is no workspace in the direction that
we want to go, show a bump animation.