swipeTracker: Allow to specify a bump range
Right now, given the snap points [0, 1, 2] we can't go below 0 or above 2
during the swipe update. In order to give more feedback during the swipe
sometimes it is useful to move the content even when we already reached
the first or the latest snap point. Address this by adding a bump range.
With a bump range specified, the gesture progress update is not limited
by the snap points boundaries. For example, with snap points [0, 1, 2]
and a bumpRange 0.1, the update swipe progress value can now be between
-0.1 and 2.1.