signalTracker: Add support for once-invoked signal handlers
In various situations we need to connect to an object signal waiting for it
to happen and immediately disconnect from it.

This can be quite annoying to do using the signalTracker without using an
extra transient signal holder to be used as tracker, while we can easily
automatize this wrapping the callback function to also perform the cleanup
operations once the signal callback has been received.

Adding support for this, using a custom GObject.ConnectFlags flag for now,
while some custom code can be removed in the GObject case when GLib will
support this natively.

This is also a pre-requisite work to have promise-based signal connections
that automatically disconnect when the promises are resolved.

Comes with tests.