tests/run-tests.sh: Use G_DEBUG=fatal-warnings by default
While the JS suite fails on exceptions that are thrown by code that is
directly called from javascript calls, it may not happen if that's called
from C code, as it's in the case of native signal callbacks.

In such cases in fact we do get a javascript error, but that's not enough
to stop the tests execution.

To ensure this, do not allow any GLib warnings or critical errors to happen
in tests without being explicitly handled (e.g. GLib.test_expect_message).