signals, signalTracker: Add Signals.DestroyableEventEmitter
Objects using JS signals could be used safely as signalTracker tracked
objects for destruction, but so far we had not a clear way to know if one
of such objects is actually destroyable, because such objects can emit any
signal without previous declaration.

To make this clear, introduce Signals.DestroyableEventEmitter class that
ensures that a such class has a destroy method that emits a 'destroy'
signal.

As per this, we can now accept such classes as input for
registerDestroyableType().