signalTracker: Avoid repeated signalData lookup on track
When tracking ab object we're looking up the object signal data twice: one
to try to connect to the destroy handler, and another one to push the
handled signal IDs.

We can avoid this by simply passing the signalData to the private
_trackDestroy() function, and by doing that we can also avoid to perform an
extra iteration on destroyableTypes if we've already a destroy handler.