signalTracker: Ensure that destroy-after signals happens last
A SignalTracker is tracking the 'destroy' signal for its owner when possible
and it does it using the "after" flag to ensure that we don't stop other
handlers to be reached.

However, if the owner is also a destroyable object, others can connect to
its own 'destroy' signal and use the AFTER flag too, and in this case we
wouldn't ever call such handler, because the SignalTracker would be cleared
earlier than this.

To ensure that this won't happen, when we're explicitly connecting to a
'destroy' signal, using the AFTER flag, force a SignalTracker re-connection
to the 'destroy' signal, so that we can be sure that no other handler will
be shadowed by it.