TransientSignalHolder: Connect to after destroy owner signal
Any object monitoring another one to perform destroy actions should try not
to react too early, we do it in all the cases by explicitly using
'connect_after' but we don't in the TransientSignalHolder.

This would have not previously worked properly in some scenarios (such as
when something else was connected to the owner 'destroy' object before
initializing the TransientSignalHolder), but as per previous commit, we can
now safely do it.

The 'TransientSignalHolder owner destruction keeps early monitored
destructions' test is keep validating this behavior.