signalTracker: Implement TransientSignalHolder via JSObject
This object can be used as tracker for other objects connections, and it was
using a GObject as base because it was the only possible way register it as
a destroyableType.
However, since the introduction of Signals.DestroyableEventEmitter this is
not true anymore, and we can avoid using a GObject (with all the logic it
involves and the C <--> JS overhead) for this task, when a simple subclass
of Signals.DestroyableEventEmitter can handle it.

Unfortunately, due to circular dependencies, we need to cheat about the way
we initialize it, because at import time Signals.DestroyableEventEmitter may
not be defined yet, and thus it's not possible to use it as a base class.

So, in workspaceThumnail, use SignalTracker to get TransientSignalHolder.
Since it changes during first initialization, it's just better to retrieve
it from the module, so that its definition will be updated after the first
instance is created.