popupMenu: Allow opening without a grab
There's a few cases where we want to open a popup menu without taking the
grab immediately. That's because the menu is opened by a gesture and we
don't know yet whether that gesture will succeed. If it won't succeed, we
don't want to take a grab.

The solution for this is to take the grab at the end of the gesture instead
of the beginning, so decouple showing the popupMenu from actually taking its
grab by introducing two methods for those two things.