appDisplay: Don't take grab for popup menu immediately
We want the app icon DnD gesture (to reorder the app grid) to have a chance
to grab, so don't let the popup menu take a grab right away after the
long-press gesture begins. Instead take that grab when the long-press
gesture ends and we know that the user didn't intend to drag.