screencastService: Add Gstreamer pipelines for AV1 encoding
AV1/libaom is used for videochat these days and as of Gstreamer 1.22 has
support for a `realtime` profile.
Thus let's use it if a recent version of `av1enc` (with support for the
`usage-profile` property) is available.

Prefer the `.webm` container format instead of `.mp4` or `.mkv` as it's
apparently best supported in practice and doesn't have license issues.