layout: Restart startup animation preparation if backgrounds changed
In case monitors changed while we were waiting for the backgrounds
update to be completed, are now getting a cancelled error.

So in such case, we should re-prepare the startup animation so that it
gets restarted with the proper monitors and screen geometries without
the risk of being stuck waiting for changes.