js/background: Add Multi-Display Wallpaper
Based on this issue #4929, I create a feature that allows users to apply a
different wallpaper to each monitor.

For the implementation, I changed the data type in gsetting-schemas for
uriBackground to an array of strings.

Then for the selection of image files to be applied, it will be based on
the monitor index, and if the element in the uri is only 1 or number of
monitors more than length of uriBackgrounds, it will directly read the
index 0 and the monitor index will be set to 0.

Signed-off-by: Taufiq Pohan <m.taufiq30s@hotmail.com>