style: Improve contrast of privacy, screenshot indicators in light theme
To improve the contrast situation, I've added variables for foreground
and background colors for the "destructive" color (used by the
screenshot indicator) and added a new "privacy" color, also with
foreground and background colors.

Having separate fg/bg colors allows appropriately contrasting colors to
be chosen which fit both dark and light modes. For the "destructive"
colors I'm using the colors from libadwaita:
https://gnome.pages.gitlab.gnome.org/libadwaita/doc/1.4/named-colors.html#destructive-colors
and for the "privacy" colors I've made color selections from the palette
that achieve appropriate contrast levels.

Additional variables for dark-specific colors are made available so
that the dark versions can be used on the "system" (transparent) panels
used in the overlay, etc, which use fixed dark mode. This ensure that
they do not incorrectly use the light-mode color variants.

The dark-specific colors are also used to fix the record button on the
screenshot ui, which was incorrectly using the light version of
$destructive_color even tho the panel is always dark, resulting in
reduced contrast.

This is a more complete fix for #7112 that covers both the screencast
recording and screenshare, camera sharing indicators.