overviewControls: Make workspaces in appGrid state a bit larger
Now that we tweaked the app grid to use larger icon sizes and fit nicely on
smaller screens, we can allow the workspaces in the appGrid state to use a
bit more vertical space.

This is useful on small screens so that the individual workspaces are more
easily recognizable, and also useful on large screens (1920x1080p or
larger), because there we're actually dealing with too much vertical space
that we need to fill.