workspace: Animate windows of type DESKTOP to the background on overview
Currently windowtype DESKTOP are not displayed in the overview. This MR makes a
layer of clones of all Meta.WindowType.DESKTOP, positions them correctly
and adds them to the backgroundgroup in the shell overview.

The windows of type DESKTOP therefore do not jarringly simply disappear from
view, but are now smoothly animated away to the background as the overview
progresses.

Uses the monitor size to scale DESKTOP windows correctly to background

For multiple monitor systems, sets the origin of the desktop window correctly