appDisplay: Move page indicators into container for the app grid
Right now the page indicators are positioned above the dash with a fixed
vertical distance of "spacing" as defined in ControlsManagerLayout.

There can be quite a big distance between dash and app grid on large screens
or certain aspect ratios though, and we want the page indicators to be
aligned right below the app grid, not right above the dash.

As a first step towards that, move the page indicators into the
scrollContainer that we use to allocate everything app grid related. To
actually align the indicator closer below the app grid, more complex changes
to the layout managers of the iconGrid will be necessary. We'll do those at
some later point, but for now this change seems nice to have anyway.