st/entry: Use hint-actor as label-actor if possible
The hint-actor is used as a placeholder to explain the purpose
of an entry, but it currently isn't exposed to accessibility.

There is no good reason to not do that, so also use the actor
as labelled-by relationship, provided the actor is a label and
no label-actor was set before.