workspace: Use existing code to add windows to workspaces
We already have code to add existing windows, and react to added/removed windows.
Hook into this code to add DESKTOP windows to background. This will initialize
the overview properly if a desktop window comes into existence or
is added while the overview is up.

Monitor all windows for change in two properties - skip-taskbar and
window_type === Meta.WindowType.DESKTOP.

Move from desktop layer in background to overview and vice-versa
if the properties change appropriately. This insures if the waylandclient
method make_desktop() is called later after the window is created, to move
the window correctly into the background, analogous to monitoring the
skip-taskbar property.

Do not intialize desktoplayer in background if there are no windows of
type DESKTOP. This avoids initialization and usage the new
backgroundLayoutManager, therefore avoiding expensive context switches with
C code in the normal layout manager used.