workspace: Rename methods for clones to diffrentiate from windows
This renames the methods and reformats code so that methods dealing
with windows are clearly separated from methods dealing with the
clones/animatedclones of the windows in the workspace. This makes
the code more intelligble.

Separate and analogous methods for window management - addition/deletion,
clone management - addition/deletiond and property signal monitoring,
instead of lumping code into one method do all the above.