messageTray: Add object holding sound for a notification
A notification can have two ways to specify a sound, via a name of a sound
in the sound theme and/or a file. Therefore introduce an object that can
old both properties and creates an abstraction over the different
source. This reduces the number of properties for a notification, which
will make it simpler to port it to GObject properties.