FdoNotificationDaemon: Actually associate a Source with an app
Actually link a notification source with an app instead of just to it's
app name and PID, which in may case don't really identify an app. E.g. for
portal applications the PID points to the xdg-desktop-portal.
Use the app when ever possible but keep using the app name and PID as a
fallback.