accessDialog: Wait for focus changes for 2 seconds
When the Access portal code is executed, it performs a routine check
to see if the app originating the request is the focused app. This check
is racy, though, especially when the app has just been opened.

Give it some leeway by waiting for at most 2 seconds after the portal
request is triggered. If an app with that app id is focused during that
time, show the portal dialog; otherwise, deny the request.