searchController: Maintain focus in the search entry when it has no text
In some cases, the first character cannot be entered in the search entry
using certain input methods, such as ibus-bamboo [1] and ibus-hiragana [2].
These input methods utilize ibus_engine_delete_surrounding_text() and
ibus_engine_commit_text() as a pair to replace character(s) before the
cursor with another character(s).

For instance, to enter 'Đ' with ibus-bamboo in the 'Surrounding Text' mode,
the user types 'Dd'. When the second 'd' is typed, ibus-bamboo removes the
previous 'D' with ibus_engine_delete_surrounding_text(), and then inserts
'Đ' with ibus_engine_commit_text(). The search entry becomes empty during
this process for a moment.

The key focus is currently lost when the letter 'D' is removed by executing
the following steps:

 1. _onTextChanged() via the 'text-changed' signal
 2. _searchCancelled()
 3. _setSearchActive(false)
 4. ControlsManager._onSearchChanged() via the 'search-active' signal
 5. ease()
 6. vfunc_unmap()
 7. reset()
 8. global.stage.set_key_focus(null);

However, ibus-bamboo still needs to commit the character 'Đ', and it fails
with the following assertion:

gnome-shell[3372]: clutter_input_focus_set_surrounding: assertion 'clutter_input_focus_is_focused (focus)' failed

This issue is almost identical to what had been discussed in [3]. Commit
35c85d9faca4a5906ba583c5a3f60201813607c6 addressed this, but it seems to
have been ineffective due to a side effect of [4].

This fix ensures that the focus is not lost when changing the search text.

[1] https://github.com/BambooEngine/ibus-bamboo
[2] https://github.com/esrille/ibus-hiragana
[3] https://bugzilla.gnome.org/show_bug.cgi?id=636341
[4] https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/205

Closes: https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/7592