workspaceAnimation: Bind clones opacity to the window actors opacity
During workspace animation we are creating clones, but they are designed
to have full opacity and not to follow the cloned actor opacity
even if the source actor use different opacity levels.

This implies that if something (e.g. an extension) changes sets a window
opacity to a different level, we don't respect that during the workspace
animation.

So, let's just bind the opacity property to the clone during the
animation. There's no need to manually unbind it since we always destroy
the clone explicitly, breaking the bind.