workspace: Sync and animate clones opacity
If windows are not fully-opaque we are currently just drawing them as
they are opaque during workspace animation and also in the app-grid
representation.

However while in the window picker view we may want to always show the
window fully opaque, the app-grid workspace view should respect their
opacity.

So, in general just always sync the clone opacity with the actual window
actor opacity, if we're instead animating to/from the window picker
increase/decrease the clone opacity according to the progress status