windowManager: Do not cancel workspace animation on kill-switch-workspace
The kill-switch-workspace event is emitted by the shell plugin (via
mutter plugin manager) just before a workspace switch is requested
however this implies that we reset all the animation state information
even if the switch is happening while another workspace switch is
happening, and this helps to contribute breaking the workspace
animations because once a new workspace switch is requested, we end up
starting over the animation from the initial point.

We now still complete the animation, but we rely on the animation
completion for that to happen now.