workspaceAnimation: Add support for adding workspace groups dynamically
We may need to include more workspaces we animate to if an user
requires a workspace switch while we're already animating between a
fixed number of workspaces.

To do this, we just recreate the monitors group container, trying to
recycle the groups we already have and adjusting the container position
so that the animation can continue were we've left it.