windowManager: Honor animation blocking when activating a workspace
When the workspace switch with move window is started we the workspace
animation is started twice: first because the workspace is activated as
part of insertWorkspace(), and then because of the explicit change
workspace in actionMoveWindow().

However, the code in insertWorkspace() was supposed not to trigger any
animation, but during the revert of commit ff9bb539 (happened in commit
3d39b32a), reverting 1cac7b22 was forgotten, leading to this issue.

This implies that the animation isn't performed properly (e.g. the
window moves to the new workspace after the animation is completed, and
it's not moved with the animation).

As per this, add back the check on blockAnimations in shouldAnimate.