workspaceAnimation: Don't change monitorGroup animation progress when animating
If a new workspace animation is started when anther was still ongoing,
we were resetting the animation progress to the initial state, causing
a visual imperfection: the animation is started again from the beginning
making the windows to be shown twice.

So, in case a new animation has been requested while we're already
animating, just keep the animation progress to the value we are at, but
ensure that the monitor group contains all the workspaces items we're
animating from and to, so that they can be dynamically added if we're
switching to a workspace that wasn't previously part of the animation
(as it happens when hitting the key-bindings again while we're switching
workspace).

As per this, also reduce the animation time so that it's proportional
to the time needed to reach the requested progress level.