FdoNotificationDaemon: Don't destroy all non-resident on app activation
Don't close all non-resident notifications if one notification of an app
is activated with the default action. This behavior isn't part of the
fdo notification specs and IMO it may confuse app developers and user
about why notifications are removed.

Additionally, with grouping by app, add in a future commit it's much
easier to dismiss all notifications of an app.